[proto] Replace enums with int32
int32 allows to add values to fields, while enums does not

+++ src/masa/sword/swapi/commands_sim.go
@@ -25,7 +25,7 @@ func getUnitMagicActionAck(msg *sword.SimToClient_Content) (*sword.UnitMagicActi
 		return nil, 0, unexpected(msg)
 	}
 	code := ack.GetErrorCode()
-	if code == sword.UnitActionAck_no_error {
+	if code == int32(sword.UnitActionAck_no_error) {
 		id := ack.GetUnit().GetId()
 		return ack, id, nil
 	}
